feat: :sparkles: Beam Intensity & Detector Exposure (#54)
Add in naive implementation of flux. This does take into account distance falloffs and cm2 detector scaling, and from experimental results matches dosage quite closely (barring internal tube filtering). However it does not account for falloff when using filtration, and especially deviates at lower voltages with high filtration.

This is a super basic model for now, comparative noise models for the DTHE are in the works, and will be implemented as part of Digital Twin support.